tree based variant generation; tests for variant tree, transpose
the previous version of dabby95() did not account for the tree
structure of streams Stream->Part->Voice. This has been refactored to
allow the chaotic mapping to work on the same voice *theoretically*
across all parts contained in a stream.

To validate these assumptions, a number of tests have been written to
see how music21 handles variants. Also tests have been written to
validate assumptions about transposition works, which is assumed the
best way to apply the chaotic mapping to chords without
splitting.. This is where this work diverges from the written text of
Dabby '95