modifications to address incorrect chaos mapping; still incorrect
somehow, the pairs of chaotic indices are not matching the reference
text (Dabby '95).
Also, I have not found a mention in the text as to handling values
that are out of the bounds of the original pitch sequence... I think
the simplest way to address this is using modulo arithmetic.